Add a comment why parallel execution is not faster
in this particular case: substring.pl takes the most time. For the old
spelling however it is very quick. In other words if you let the snail
and the rabbit race in parallel it will not be much faster than if
they run in sequence. You basically still just wait for the snail.