Reset the form after adding a new pattern
to mimic the behavior or a regular form. To do that
add-hyphenation-pattern! was extended so that a callback for success
and error can be passed in. Save and update get different callbacks.
Save does the resetting and re-loading of the patterns. Update only
updates the exiting (in-memory) patterns (after having successfully
saved it on the server).