Always hyphenate against the base dictionary
This makes the GUI more consistent. If you hyphenate using the
exported dic then seemingly the existing hyphenations are correct, so
there is no need to even have them in the database as exceptions.

This also removes the need for a stateful hyphenator. No more atom to
keep the Hyphenators and no more reloading after an export.